deployed the app for production level
